requirements.txt
fastmcp[fastapi]>=0.3.2
uvicorn[standard]>=0.29
httpx[http2]>=0.27
jsonschema>=4.22
python-jose>=3.3
pyyaml>=6.0
orjson>=3.10
fastapi-limiter>=0.1
respx>=0.20  # para tests
pytest>=7.0
pytest-asyncio>=0.21